[ERMP]:Refactoring, Combining edit and add forms.
