+++ src/components/MovieBaseForm/README.md
@@ -1,0 +1,1 @@
+# Base movie form component
