+++ src/components/MovieBaseForm/index.js
@@ -1,0 +1,3 @@
+import MovieBaseForm from './js/MovieBaseForm';
+
+export default MovieBaseForm;
