+++ src/components/MultiSelectInput/js/MultiSelectInput.jsx
@@ -9,7 +9,7 @@ const ArrowRenderer = ({ expanded }) => (expanded ?
   <img className="arrow" src={Constants.ARROWS[0]} alt="arrow_up"/> :
   <img className="arrow" src={Constants.ARROWS[1]} alt="arrow_down"/>);
 
-const buildValues = (elems) => elems.map((e) => ({ label: e, value: e }));
+const buildValues = (elements) => elements.map((e) => ({ label: e, value: e }));
 
 export default function MultiSelectInput({ title, options, preselected, onAction, onBlur, error }) {
 
@@ -37,8 +37,6 @@ export default function MultiSelectInput({ title, options, preselected, onAction
 
 ArrowRenderer.propTypes = {
   expanded: PropTypes.bool.isRequired,
-  error: PropTypes.string,
-  onBlur: PropTypes.func,
 };
 
 MultiSelectInput.propTypes = {
@@ -50,8 +48,12 @@ MultiSelectInput.propTypes = {
   ).isRequired,
   preselected: PropTypes.arrayOf(PropTypes.string),
   onAction: PropTypes.func.isRequired,
+  error: PropTypes.string,
+  onBlur: PropTypes.func,
 };
 
 MultiSelectInput.defaultProps = {
   preselected: [],
+  error: '',
+  onBlur: '',
 };
